Treat 'missing trie node errors' for deeper blocks
The algorithm tries to fetch the state from the specific block of last event
for better accuracy. Depending on the actual block depth and node configuration
(sync mode and garbage collector), the node may return the error {'code': -32000,
'message': 'missing trie node ...'}. In this case we print the warning and take the
latest available state.